added regex verification before INSERT for email and password

+++ controllers/password-controller.php
@@ -7,26 +7,29 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
         $password = $_POST["password"];
         $token = $_POST["token"];
         $stmt = $_DB->execute("SELECT email, token FROM reset_passwd WHERE token = :token", ["token" => $token])->fetch();
-        if ( isset($stmt) and $stmt["token"] == $token) {
-            $email = $stmt["email"];
-            // Generate hashed password using PASSWORD_ARGON2ID algorithm
-            $options = [
-                "memory_cost" => PASSWORD_ARGON2_DEFAULT_MEMORY_COST,
-                "time_cost" => PASSWORD_ARGON2_DEFAULT_TIME_COST,
-                "threads" => PASSWORD_ARGON2_DEFAULT_THREADS,
-            ];
-            $hashedPassword = password_hash(
-                $password,
-                PASSWORD_ARGON2ID,
-                $options
-            );
+        if (isset($stmt) and $stmt["token"] == $token) {
+            $regex = "/^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$/";
+            if (preg_match($regex, $password)) {
+                $email = $stmt["email"];
+                // Generate hashed password using PASSWORD_ARGON2ID algorithm
+                $options = [
+                    "memory_cost" => PASSWORD_ARGON2_DEFAULT_MEMORY_COST,
+                    "time_cost" => PASSWORD_ARGON2_DEFAULT_TIME_COST,
+                    "threads" => PASSWORD_ARGON2_DEFAULT_THREADS,
+                ];
+                $hashedPassword = password_hash(
+                    $password,
+                    PASSWORD_ARGON2ID,
+                    $options
+                );
 
-            $query = "UPDATE users SET password = :new_password WHERE email = :email";
-            $results = $_DB
-                ->execute($query, ["new_password" => $hashedPassword, "email" => $email]);
-            $results = $_DB
-                ->execute("DELETE FROM reset_passwd WHERE email = :email",["email" => $email]);
-            echo "success";
+                $query = "UPDATE users SET password = :new_password WHERE email = :email";
+                $results = $_DB
+                    ->execute($query, ["new_password" => $hashedPassword, "email" => $email]);
+                $results = $_DB
+                    ->execute("DELETE FROM reset_passwd WHERE email = :email", ["email" => $email]);
+                echo "success";
+            }
         } else {
             echo "invalid_token";
         }
